ethernet: fix for bug #351: Ethernet PAUSE frames are dropped
fills destination address, checks special multicast pause address at incoming packets

Bug description:
Ethernet PAUSE frames are generated in EtherEncap, EtherLLC and MACRelayUnitBase classes.
However the destination address is not filled in, so the receiving MACs drop the frame,
because it is not destined to them (unless in promiscous mode).
This makes the flow control ineffective.

The destination address of the frame may be set to either the unique DA of the station to be paused,
or to the globally assigned multicast address 01-80-C2-00-00-01 (hex). This multicast address has been
reserved by the IEEE 802.3 standard for use in MAC Control PAUSE frames. It is also reserved in the
IEEE 802.1D bridging standard as an address that will not be forward by bridges. This ensures the frame
will not propagate beyond the local link segment.

X-Committed-on-Branch: topic/etherfixes