ethernet: EthernetInterface: move queue module to between EtherEncap and EtherMAC modules
Reason: this way the queue will contain Ethernet frames, so for example this way the
queue can ensure that PAUSE frames overtake normal frames. (EtherEncap can send PAUSE
frames by command from the upper layer).

From now on, the queue module in EthernetInterface needs a classifier class that
works on Ethernet frames (not on IPv4 datagrams for example). This commit implements
EtherFrameClassifier and introduces it. It ensures that PAUSE frames have higher priority.

The queue module is conditional: when queueType is "NoQueue", there will be
no queue module, and the MAC will its use internal queue.