fixes for bug #190 by Alfonso Ariza: 802.11 STA nodes abort the simulation
If the STA node is not associated to an AP it will send a packet with a invalid address to the mac and the mac will stop the simulation.
It's necessary to modify the files Ieee80211STA.cc and Ieee80211STASimple.cc

Copied changes from bugreport and from inetmanet.

In Ieee80211MgmtSTA, I removed the old solution, and inserted new solution from manet.