IPv4 bugfix for #459: there is no fragmentation on the loopback interface
Solution: drop packet and send ICMP if DontFragment bit of packet set and
packet length bigger than MTU

bug was:
If a packet is transmitted through the loopback interface, then is
delivered directly, and the fragmentation method is not called.
The loopback interface has an MTU, that should be respected by the
IPv4 module.