throw cRuntimeError adjustments: "this" as 1st arg is not needed when thrown from module
The error message produced by cRuntimeError will in any case contain the
module (fullpath+type) which was active when the exception was thrown,
whether "this" is there or not.

The first, "this" arg (a cObject*) can be supplied to provide additional
info into the error message, but it will be ignored if it's the same
as the active module because it brings no extra info (see above).