util: added PatternMatcher
This is a pattern matcher class for inifile-style patterns.

It's near copy of an OMNeT++ internal class (in src/common/).
Since we'd like to use patterns in INET (and this is generally
useful functionality), we'll make part of the public OMNeT++ API
in the 4.3 version; until then we have this copy in inet/src/util/.