IPv4: parent interface can be NULL in IPv4MulticastRoute
meaning that the parent interface is the shortest path interface to the source (according to the routing table).

This can be used when the route entry matches to any source (i.e. origin is 0.0.0.0/0).

These routing entries enables us to build a source independent routing trees,
and to be backward compatible with the previous multicast routing algorithm.