IPv4Route/IPv6Route: add administrative distance field
The route entries are sorted first by administrative distance, then by metric.
When looking up the best route for a destination, then routes with lower
administrative distances are preferred. Metrics are compared only if
the administrative distances are the same.

Default values for these distances are defined in RouteAdminDist enums,
but they are set only for interface and static routes.