build: using the new opp_featuretool
The opp_featuretool now generates a header file (src/inet/features.h) that is
included in the INETDefs.h file. It contains all the WITH_FEATURE defines
that are required to compile the project with the current feature
enablements. This is good, because other projects inet can safely include
any include files from inet (as long as the INETDefs.h file is included).