Ieee80211Mac: added a new IEEE 802.11 MAC implementation.
This is the new IEEE 802.11 MAC implementation. The implementation is entirely based on the standard IEEE 802.11™-2012 Part 11: Wireless LAN Medium Access Control (MAC) and Physical Layer (PHY) Specifications.

It is designed to be a flexible and modular architecture, therefore it allows the users to replace the main components with their own implementations. This design facilitates experimenting with new protocols within the MAC layer.

Due to its modular design the implementation separates the following components (i.e. segregates them in their own classes):

- coordination functions
- channel access methods
- MAC data services
- frame exchange sequences
- queues
- TX/RX components

It includes several enhancements and new features to the previous version such as block acknowledgement and MSDU aggregation.