Common: Added a new shared ptr implementation called IntrusivePtr and support for using it instead of std::shared_ptr.
The new implementation uses only 1 counter (no weak counter) that is directly
added to the shared object using inheritance. This is the reason why it's called
intrusive. This shared pointer is not thread safe, so it doesn't use atomic
operations to increment/decrement counters, but it's not a requirement for
simulation. According to speed measurements (see tests folder) the release builds
run 5-10% faster.