Fixes infinite looping in GPSR which led to zeno simulation behavior
If packets are sent by a node that has to enter perimeter routing (because the destination node is not in the neighborhood table (due to lost GPSR beacons or placement out of transmission range)), the loop trying to find the next neighbor errorneously did an intersection of the start_node<->destination_node line and the start_node<->next_neighbor line.
This resulted in a situation where an intersection was always given, so that the loop was never left.

This commit introduces checks to prevent intersections at the perimeter start node.