fix headers modulation,
Page 2546: This definition results in a QBPSK modulation on the second symbol of VHT-SIG-A where the constellation of
the data tones is rotated by 90º counter-clockwise relative to the first symbol of VHT-SIG-A and relative to the non-HT
signal field in VHT PPDUs (Figure 21-20). In VHT PPDUs, the VHT-SIG-A is transmitted with the same number of
subcarriers and the same cyclic shifts as the preceding non-HT portion of the preamble.

Note : In simulation is the same than to use BPSK