ospf: Updated fingerprints for "ospf: adding checksum to the OSPF messages" patch.
These fingerprint tests were left out accidentally, but has been verified the same way.
The fingerprints were changed due to the following:
-    B packetLength = IPv4_MAX_HEADER_LENGTH + OSPF_LSA_HEADER_LENGTH;
+    B packetLength = IPv4_MAX_HEADER_LENGTH + OSPF_HEADER_LENGTH + B(4);

This change has been verified by removing it and re-running the fingerprint tests.
The extra 4 bytes is the number of LSAs in the update packet.

The tests can be run in tests/fingerprint with:
./fingerprinttest --debug -m ospf