CsmaCa: The length field is not optional at the moment.
If we really wanted to make it optional, then we should explicitly implement
optional header length field by introducing new parameters or parameter values.
For example, the sender should be aware of not including the length field in
the packet, the receiver should now read the length field from the packet, and
the packet should always contain an invalid length field, and the serializer
should never serialize/deserialize a header length field.