MessageDispatcher: Dropped updateTags().
Removing the DispatchProtocolReq was a bad idea due to many packet processor
components want to be transparent and it would make life complicated for them.

We must either remove the DispatchProtocolReq and put it back from protocol
transparent components, or we must remove the DispatchProtocolReq from actual
destination protocols to avoid infinite loops in the packet dispatching. We
chose the latter.