follow 'all: Hook classes converted to submodules': changed fingerprint
Only the /~tND fingerprint changed, because this simulation uses the
HostAutoConfigurator for configure the IPv4 addresses of network.
The HostAutoConfigurator calculates IPv4 addresses based on module ID
of containing networknode. This simulation generate network nodes
dynamically, so the IP address of generated networknode depends on the
number of modules already generated. The followed commit increased the
number of generated modules when the hook needed, so the IPv4 address
also changed, so the packet content changed.