Gptp: Fixed the computation of gmRateRatio.
The formula uses the previous and the current timestamp of the peer time
synchronization frame sending events. We could not find parts of the
standard that would justify this but the OneMasterClock example of the
gPTP time synchronization showcase proves this method to be accurate
and the old method to be completely wrong.