EthernetCutthrough*: Refactored to pass up cut-through packets with a StreamBufferChunk.
The old implementation directly streamed packets from one interface to the other,
the new implementation passes the incoming packet up just like normal packets.
The difference for cut-through switching is that the cut-through source doesn't
wait until the whole packet is received, but only a portion of it. The packet
contains a special chunk type for the remaining part that states that this part
is continuously being filled up.