CreditBasedGate: Credit updates on closed gate
This change adresses the issue that the credit of CreditBasedGates was increasing on  while the corresponding PeriodicGate was closed. Per IEEE 802.Q-2018 credit should be frozen on closed gates. Moreover an option was added to toggle the accumulation of credit during pre-closing time, i.e. when the PeriodicGate is open but not long enough anymore to send the next package. According to IEEE credit should still increase, which leads to higher usage of bandwidth but fairness between different queues isn't as good anymore and credit could be unbounded. Therefore a parameter allows to switch the behavior to accomodate different use cases.