.github: Switch to newer Docker image for cross builds
This should fix the failures of the Windows build, because this image
is now based on Ubuntu 22.04, which has GNU Make 4.3 in it, which
understands group targets, so it won't invoke the same linker command
twice in parallel unnecessarily to make the .dll and the .dll.a files.
Linking takes about 5 gigs of RAM in debug mode, so the CI runners
could barely fit it once, but definitely not twice at the same time.