tcp: tcp control flow implementation: using tcp socket with read() command
Until now, all TCP sockets operated in "auto-read" mode, i.e. all data
that arrived in the TCP connection were immediately sent up to the app.

This commit implements read() requests for sockets. This read mode means that
the user must send an explicit read request, similar to a real-life socket
read() call, and the TCP will respond either with already queued data, or
if there is no queued data, with new data that arrives next on the
connection. A limit for the number of bytes to be sent up ("buffer
size") can be specified in the read request.

TCP flow control is only effective in read mode. (An autoread TCP connection
will always advertise the value set in the "advertisedWindow" parameter.)
With READ mode, the advertised window will be decreased by newly arrived
data unless READ requests make room in the buffer.

Summary:
- flow control only works in READ mode (autoread always advertizes full buffer size)
- READ mode can be turned on in TcpOpenCommand, by setting autoRead=false
  (autoRead is still the default, for backward compatibility)
- a READ request is represented by TcpReadCommand, with numberOfBytes
  being an upper limit
- after a READ request, issuing a new READ request is only allowed after
  the previous one has been fulfilled
- related: sending up TCP_I_PEER_CLOSED indication after receiving a FIN
  is now delayed until all data have been read by the app