doc/users-guide: further wording suggestions generated using llm
llm instructions: "The following reStructuredText file is part of the
INET Framework for OMNeT++. At places where it addresses the user
as 'you', change it to neutral, e.g. to passive voice or 'one' as subject.
Keep all markup and line breaks intact as much as possible!
Lines starting with '..' are comments, do not touch them!"