all: Followed the recent addition of the templated get method in the units library.
The idea is that it must be explicitly stated in what unit the value is needed.
If the type of an argument or a variable is changed, it will have no effect on retrieving the numeric value in a given unit.
For example, instead of mW(power).get() you can write power.get<mW>().
This way is more intentional than it was before.