mrp: fix: do not truncate CCM interval to milliseconds
Allowed values are 10ms and 3.3ms, and msec_trunc() turns 3.3ms into 3ms.

This is a bug in the original code. trunc_msec(ccmInterval) used to be
SimTime(ccmInterval, SIMTIME_MS) but in omnetpp 5.x this SimTime ctor
takes int64_t, so the fractional part of the double was discarded.
SimTime ctor changed in omnetpp-6.0 due to being error-prone.