all: Removed all ChangeLog files because they are no longer maintained.
The decision to remove these files was made on a team meeting.

With the recent advancements of AI technologies there are already better ways
to extract more useful information from the git repository. Maintaining the
ChangeLog files took a non-trivial amount of time; not important commits had
to be filtered out, many commit messages were rephrased, more often than not
commits affect somewhat unrelated folders, etc. Also, it was unclear who were
the actual users of the ChangeLog files to begin with.