Ieee80211AgentSta: partial fix for #1022: defaultSsid parameter enhanced.
defaultSsid now accepts a list of SSIDs the node is allowed to connect to,
and chooses the one with the best signal strength from them.

If only one SSID is listed (such as in the default case, "SSID"), and
there are multiple APs with that SSID, the agent will now choose the
one with the strongest signal (instead of the first one as before).

Refactoring: chooseBSS() now encapsulates all of BSS selection functionality.