Ieee80211AgentSta: remove default value "SSID" for defaultSsid parameter
It was arbitrary. This change does not change the behavior of
example or showcase simulations, because in those simulations the AP
named "SSID" is the only AP in the network.

It however does change the network configurator behavior, because it
used defaultSsid for identifiying WLAN membership. This is addressed
in the next commit.