doc
This project is the INET Framework of OMNeT++. The task is to write a
very brief NED documentation for the NED files
that have no documentation at all.

Read existing NED files, e.g. those in src/inet/applications/ethernet/
and in src/inet/node, for the general format and style of comments.

For additional input for the docu of each module, look:

    - at the the c++ implementation files, .h and .cc, usually next to the .ned file
    - into the INET User Manual (doc/src/users-guide/) -- many chapters are relevant
    - into the showcases in the showcases/ folder that contain both simulation models and commentary in rst format

Place the doc comment DIRECTLY ABOVE the module or module interface definition ("module X...", "simple X..", or "moduleinterface X..."),
with no intervening blank line. The doc comment should begin and end with plain "//" (blank comment) lines. Break around column 80.

The list of files to process is in the file src/undocumented.txt. Use this file to tracking
progress (e.g. for each fine done, add DONE next to the filename).

Each docu should:
    - Contain just one paragraph of general description of its functionality or purpose, and NOTHING ELSE. Omit the @see line too!!!
    - Omit the phrase "This is..", "This module..." or "This module interface..." from the beginning, as it is awkward and superfluous
      in contexts (e.g. extracted online documentation) where this brief info is used.