fingerprints: updated the graphical (tyf) fingerprints
The visual results are the same. The difference of fingerprints are
coming from the previous refactorings where certain display string
changes were moved to the SimpleModule base class. So instead
of changing t tag in the mosule class, the responsibility to set the t
tag was moved to the base class. This results in changing different tags
in different other (but the end result remains the same)