build: removed checkenvir target
we cannot enforce the requirement of having setenv sourced for a build
because the user might have started the IDE from a desktop shortcut
and the build process is executed in that environment.

There is no way to properly set up an environment in that use-case. The build
process should NOT assume an environment.