steps 1-5 wip
prompt:

This is the INET project. Our goal is to work on the Protocol tutorial (tutorials/protocol folder).
This tutorial is incomplete, but the simulations for some of the steps have been completed. We need to
write documentation for the completed steps.

Before we begin, please:

- Read all documentation, ini and ned files for the existing tutorials (wireless, configurator, fingerprint, queueing) for
their overall tutorial structure (e.g. the simulations, step progression) and the structure and style of the documentation for the steps.
- Read all ned documentation and c++ code for protocol elements (src/protocolelement)
- Read the existing steps for the protocol tutorial

Some information: this tutorial is similar to the wireless and configurator in the sense that the steps
build upon each other, there is a progression. It is also similar to the queueing tutorial in the sense
that each step introduces a new functionality/module.

The steps should use the following structure:

--------
Step 1
======

Goals
-----

Short introduction

The Model
---------

Doesn't necessarily need subsections. Contains:

- the network
- description of new protocol elements added if any
- the configuration
- any explanation thats required

Results
-------

Presents what happens. Usually with a video, maybe screenshots.
--------

Please wrap the generated text at around 80 columns.