Aodv: add gateway for routing external addresses
TODO: delay the resolution of gatewayAddress?

The AODV protocol had no support for routing addresses in external
IP networks. This commit implements a simple solution for the most
basic use case: a single predefined, statically configured gateway.

The gateway address is a parameter of AODV (all nodes will know it).
When a node encounters an external address, it forwards it to the
next hop towards the gateway. AODV attaches the NextHopAddressReq
to the packet to achieve this.