Aodv: add support for routing external addresses towards a gateway
The AODV protocol had no support for routing to addresses in external
IP networks. This commit implements a simple solution for the most
basic use case: a single predefined, statically configured gateway.

The gatewayAddress is a parameter of AODV (all nodes will know it).
When a node encounters an external address, it adds a deafult (inactive)
route to the gateway and starts route dicovery for the host defined by gatewayAddress.
Once the route dicovered it forwards the queued packet according to that.

- added suport to RoutingTable to properly display AODV maintained routes
- fixed AODV to use the IP address of the interface configured for AODV
  instead of using the router ID (which might be different if multiple
  interfaces are present in the node - and also can be manually changed
  to any arbitrary ID)
- added an additional configuration to examples/aodv to demonstrate the feature.

Aodv: fixes for crash after default route was expunged