showcases: pcaprecording: fixed packetFilter for recording TCP data; fixed recording ipv4 frames;
The previous filter for TCP data was misleading, as it used the length of the IP frames, instead of TCP.
Also, it didn't record the last TCP data packet, which can be of different length.
For recording ipv4 frames, the modulename patterns is not needed, as that would record packets
at the ipv4 module. The intent is to record ipv4 packets at the network interfaces.